requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
python-dotenv
# pip install -r requirements.txt
